Tries to determine the number of littters that a container can hold. The container is passed as an array, with each number repreesenting one unit. The zeros represent a horizontal displacement. Whenever the number in the array is not zero, it means that the container has a "wall" at that position that has a height equivalent to the number.
As of now, this solution is not complete, as it fails the cases in which 
the "sub tanks" can have a valid height that is greater thn the valid 
heoight of the entire solution.

View source to understand the concept of valid height as used here.

I shall revesit this later in the evening.